Default to GHES host if only GHES is authenticated (#3286)
