+++ context/remote_test.go
@@ -58,3 +58,14 @@ func Test_translateRemotes(t *testing.T) {
 		t.Errorf("got %q", result[0].RepoName())
 	}
 }
+
+func Test_FilterByHosts(t *testing.T) {
+	r1 := &Remote{Remote: &git.Remote{Name: "mona"}, Repo: ghrepo.NewWithHost("monalisa", "myfork", "test.com")}
+	r2 := &Remote{Remote: &git.Remote{Name: "origin"}, Repo: ghrepo.NewWithHost("monalisa", "octo-cat", "example.com")}
+	r3 := &Remote{Remote: &git.Remote{Name: "upstream"}, Repo: ghrepo.New("hubot", "tools")}
+	list := Remotes{r1, r2, r3}
+	f := list.FilterByHosts([]string{"example.com", "test.com"})
+	assert.Equal(t, 2, len(f))
+	assert.Equal(t, r1, f[0])
+	assert.Equal(t, r2, f[1])
+}
