+++ internal/config/stub.go
@@ -49,3 +49,11 @@ func (c ConfigStub) Write() error {
 	c["_written"] = "true"
 	return nil
 }
+
+func (c ConfigStub) DefaultHost() (string, error) {
+	return "", nil
+}
+
+func (c ConfigStub) DefaultHostWithSource() (string, string, error) {
+	return "", "", nil
+}
