+++ internal/ghinstance/host.go
@@ -8,25 +8,9 @@ import (
 
 const defaultHostname = "github.com"
 
-var hostnameOverride string
-
 // Default returns the host name of the default GitHub instance
 func Default() string {
 	return defaultHostname
-}
-
-// OverridableDefault is like Default, except it is overridable by the GH_HOST environment variable
-func OverridableDefault() string {
-	if hostnameOverride != "" {
-		return hostnameOverride
-	}
-	return defaultHostname
-}
-
-// OverrideDefault overrides the value returned from OverridableDefault. This should only ever be
-// called from the main runtime path, not tests.
-func OverrideDefault(newhost string) {
-	hostnameOverride = newhost
 }
 
 // IsEnterprise reports whether a non-normalized host name looks like a GHE instance
