+++ pkg/cmd/factory/remote_resolver.go
@@ -4,13 +4,15 @@ import (
 	"errors"
 	"net/url"
 	"sort"
-	"strings"
 
 	"github.com/cli/cli/context"
 	"github.com/cli/cli/git"
 	"github.com/cli/cli/internal/config"
 	"github.com/cli/cli/internal/ghinstance"
+	"github.com/cli/cli/pkg/set"
 )
+
+const GH_HOST = "GH_HOST"
 
 type remoteResolver struct {
 	readRemotes   func() (git.RemoteSet, error)
@@ -18,7 +20,7 @@ type remoteResolver struct {
 	urlTranslator func(*url.URL) *url.URL
 }
 
-func (rr *remoteResolver) Resolver(hostOverride string) func() (context.Remotes, error) {
+func (rr *remoteResolver) Resolver() func() (context.Remotes, error) {
 	var cachedRemotes context.Remotes
 	var remotesError error
 
@@ -48,50 +50,45 @@ func (rr *remoteResolver) Resolver(hostOverride string) func() (context.Remotes,
 			return nil, err
 		}
 
-		knownHosts := map[string]bool{}
-		knownHosts[ghinstance.Default()] = true
-		if authenticatedHosts, err := cfg.Hosts(); err == nil {
-			for _, h := range authenticatedHosts {
-				knownHosts[h] = true
-			}
+		authedHosts, err := cfg.Hosts()
+		if err != nil {
+			return nil, err
 		}
+		defaultHost, src, err := cfg.DefaultHostWithSource()
+		if err != nil {
+			return nil, err
+		}
+		// Use set to dedupe list of hosts
+		hostsSet := set.NewStringSet()
+		hostsSet.AddValues(authedHosts)
+		hostsSet.AddValues([]string{defaultHost, ghinstance.Default()})
+		hosts := hostsSet.ToSlice()
 
-		// filter remotes to only those sharing a single, known hostname
-		var hostname string
-		cachedRemotes = context.Remotes{}
+		// Sort remotes
 		sort.Sort(resolvedRemotes)
 
-		if hostOverride != "" {
-			for _, r := range resolvedRemotes {
-				if strings.EqualFold(r.RepoHost(), hostOverride) {
-					cachedRemotes = append(cachedRemotes, r)
-				}
-			}
+		// Filter remotes by hosts
+		cachedRemotes := resolvedRemotes.FilterByHosts(hosts)
 
-			if len(cachedRemotes) == 0 {
-				remotesError = errors.New("none of the git remotes configured for this repository correspond to the GH_HOST environment variable. Try adding a matching remote or unsetting the variable.")
-				return nil, remotesError
+		// Filter again by default host if one is set
+		// For config file default host fallback to cachedRemotes if none match
+		// For enviornment default host (GH_HOST) do not fallback to cachedRemotes if none match
+		if src != "" {
+			filteredRemotes := cachedRemotes.FilterByHosts([]string{defaultHost})
+			if src == GH_HOST || len(filteredRemotes) > 0 {
+				cachedRemotes = filteredRemotes
 			}
-
-			return cachedRemotes, nil
-		}
-
-		for _, r := range resolvedRemotes {
-			if hostname == "" {
-				if !knownHosts[r.RepoHost()] {
-					continue
-				}
-				hostname = r.RepoHost()
-			} else if r.RepoHost() != hostname {
-				continue
-			}
-			cachedRemotes = append(cachedRemotes, r)
 		}
 
 		if len(cachedRemotes) == 0 {
-			remotesError = errors.New("none of the git remotes configured for this repository point to a known GitHub host. To tell gh about a new GitHub host, please use `gh auth login`")
+			if src == GH_HOST {
+				remotesError = errors.New("none of the git remotes configured for this repository correspond to the GH_HOST environment variable. Try adding a matching remote or unsetting the variable.")
+			} else {
+				remotesError = errors.New("none of the git remotes configured for this repository point to a known GitHub host. To tell gh about a new GitHub host, please use `gh auth login`")
+			}
 			return nil, remotesError
 		}
+
 		return cachedRemotes, nil
 	}
 }
