+++ pkg/cmd/secret/list/list.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/cli/cli/api"
-	"github.com/cli/cli/internal/ghinstance"
+	"github.com/cli/cli/internal/config"
 	"github.com/cli/cli/internal/ghrepo"
 	"github.com/cli/cli/pkg/cmd/secret/shared"
 	"github.com/cli/cli/pkg/cmdutil"
@@ -20,6 +20,7 @@ import (
 type ListOptions struct {
 	HttpClient func() (*http.Client, error)
 	IO         *iostreams.IOStreams
+	Config     func() (config.Config, error)
 	BaseRepo   func() (ghrepo.Interface, error)
 
 	OrgName string
@@ -28,6 +29,7 @@ type ListOptions struct {
 func NewCmdList(f *cmdutil.Factory, runF func(*ListOptions) error) *cobra.Command {
 	opts := &ListOptions{
 		IO:         f.IOStreams,
+		Config:     f.Config,
 		HttpClient: f.HttpClient,
 	}
 
@@ -74,7 +76,20 @@ func listRun(opts *ListOptions) error {
 	if orgName == "" {
 		secrets, err = getRepoSecrets(client, baseRepo)
 	} else {
-		secrets, err = getOrgSecrets(client, orgName)
+		var cfg config.Config
+		var host string
+
+		cfg, err = opts.Config()
+		if err != nil {
+			return err
+		}
+
+		host, err = cfg.DefaultHost()
+		if err != nil {
+			return err
+		}
+
+		secrets, err = getOrgSecrets(client, host, orgName)
 	}
 
 	if err != nil {
@@ -131,8 +146,7 @@ func fmtVisibility(s Secret) string {
 	return ""
 }
 
-func getOrgSecrets(client *api.Client, orgName string) ([]*Secret, error) {
-	host := ghinstance.OverridableDefault()
+func getOrgSecrets(client *api.Client, host, orgName string) ([]*Secret, error) {
 	secrets, err := getSecrets(client, host, fmt.Sprintf("orgs/%s/actions/secrets", orgName))
 	if err != nil {
 		return nil, err
