+++ pkg/cmd/secret/set/http.go
@@ -9,7 +9,6 @@ import (
 	"strings"
 
 	"github.com/cli/cli/api"
-	"github.com/cli/cli/internal/ghinstance"
 	"github.com/cli/cli/internal/ghrepo"
 	"github.com/cli/cli/pkg/cmd/secret/shared"
 )
@@ -47,8 +46,7 @@ func getPubKey(client *api.Client, host, path string) (*PubKey, error) {
 	return &pk, nil
 }
 
-func getOrgPublicKey(client *api.Client, orgName string) (*PubKey, error) {
-	host := ghinstance.OverridableDefault()
+func getOrgPublicKey(client *api.Client, host, orgName string) (*PubKey, error) {
 	return getPubKey(client, host, fmt.Sprintf("orgs/%s/actions/secrets/public-key", orgName))
 }
 
@@ -67,11 +65,10 @@ func putSecret(client *api.Client, host, path string, payload SecretPayload) err
 	return client.REST(host, "PUT", path, requestBody, nil)
 }
 
-func putOrgSecret(client *api.Client, pk *PubKey, opts SetOptions, eValue string) error {
+func putOrgSecret(client *api.Client, host string, pk *PubKey, opts SetOptions, eValue string) error {
 	secretName := opts.SecretName
 	orgName := opts.OrgName
 	visibility := opts.Visibility
-	host := ghinstance.OverridableDefault()
 
 	var repositoryIDs []int
 	var err error
