+++ pkg/cmd/ssh-key/list/http.go
@@ -20,12 +20,12 @@ type sshKey struct {
 	CreatedAt time.Time `json:"created_at"`
 }
 
-func userKeys(httpClient *http.Client, userHandle string) ([]sshKey, error) {
+func userKeys(httpClient *http.Client, host, userHandle string) ([]sshKey, error) {
 	resource := "user/keys"
 	if userHandle != "" {
 		resource = fmt.Sprintf("users/%s/keys", userHandle)
 	}
-	url := fmt.Sprintf("%s%s?per_page=%d", ghinstance.RESTPrefix(ghinstance.OverridableDefault()), resource, 100)
+	url := fmt.Sprintf("%s%s?per_page=%d", ghinstance.RESTPrefix(host), resource, 100)
 	req, err := http.NewRequest("GET", url, nil)
 	if err != nil {
 		return nil, err
