+++ pkg/cmd/ssh-key/list/list.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/cli/cli/internal/config"
 	"github.com/cli/cli/pkg/cmdutil"
 	"github.com/cli/cli/pkg/iostreams"
 	"github.com/cli/cli/utils"
@@ -14,13 +15,15 @@ import (
 
 type ListOptions struct {
 	IO         *iostreams.IOStreams
+	Config     func() (config.Config, error)
 	HTTPClient func() (*http.Client, error)
 }
 
 func NewCmdList(f *cmdutil.Factory, runF func(*ListOptions) error) *cobra.Command {
 	opts := &ListOptions{
-		HTTPClient: f.HttpClient,
 		IO:         f.IOStreams,
+		Config:     f.Config,
+		HTTPClient: f.HttpClient,
 	}
 
 	cmd := &cobra.Command{
@@ -44,7 +47,17 @@ func listRun(opts *ListOptions) error {
 		return err
 	}
 
-	sshKeys, err := userKeys(apiClient, "")
+	cfg, err := opts.Config()
+	if err != nil {
+		return err
+	}
+
+	host, err := cfg.DefaultHost()
+	if err != nil {
+		return err
+	}
+
+	sshKeys, err := userKeys(apiClient, host, "")
 	if err != nil {
 		if errors.Is(err, scopesError) {
 			cs := opts.IO.ColorScheme()
